Add better Error handing for redux form
I was getting an error from redux forms that stated : 'Asynchronous validation promise was rejected without errors.'

This is not very descriptive, as the problem was that I actually forgot to launch the contracts on ganache. 

Adding the lines as seen in the commit allows you to catch the error from TickerRegistry.getDetails(v) if one exists.